♻️ refactor: 상위 페이지에서 댓글상태 관리하도록 수정
