+++ src/app/community/[boardType]/[_id]/CommentList.tsx
@@ -1,27 +1,8 @@
 'use client';
-import { useEffect, useState } from 'react';
-import { getReplies } from '@/data/functions/post';
 import CommentItem from './CommentItem';
 import { PostReply } from '@/types';
 
-export default function CommentList({ _id }: { _id: number }) {
-  const [replies, setReplies] = useState<PostReply[]>([]);
-  const [error, setError] = useState<string | null>(null);
-
-  useEffect(() => {
-    const fetchReplies = async () => {
-      const res = await getReplies(_id);
-      if (res.ok) {
-        setReplies(res.item);
-      } else {
-        setError(res.message);
-      }
-    };
-
-    fetchReplies();
-  }, [_id]);
-
-  if (error) return <p>{error}</p>;
+export default function CommentList({ replies }: { replies: PostReply[]; }) {
 
   return (
     <div className="mt-5">
